Add email sending functionality

+++ unsubscribe.php
@@ -8,22 +8,38 @@ require __DIR__ . "/required.php";
 
 require __DIR__ . "/lib/iputils.php";
 
-$address = $VARS['a'];
+?>
+<!DOCTYPE html>
+<title>Unsubscribe</title>
+<meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+<?php
 
-engageRateLimit();
+if (isset($VARS['a'])) {
+    echo "<p>";
+    $address = $VARS['a'];
 
-if (!filter_var($address, FILTER_VALIDATE_EMAIL)) {
-    die("Invalid email address.");
-}
+    engageRateLimit();
 
-$address = str_replace("%", '\%', $address);
+    if (!filter_var($address, FILTER_VALIDATE_EMAIL)) {
+        die("Invalid email address.");
+    }
 
-echo $address;
+    $address = str_replace("%", '\%', $address);
 
-if ($database->has('addresses', ['email' => $address])) {
-    $count = $database->count('addresses', ['email' => $address]);
-    $database->delete('addresses', ['email' => $address]);
-    die("$address has been removed from $count mailing " . ($count === 1 ? "list" : "lists") . ".");
+    if ($database->has('addresses', ['email' => $address])) {
+        $count = $database->count('addresses', ['email' => $address]);
+        $database->delete('addresses', ['email' => $address]);
+        die("$address has been removed from $count mailing " . ($count === 1 ? "list" : "lists") . ".");
+    } else {
+        die("$address has already been removed.");
+    }
 } else {
-    die("$address has already been removed.");
+    ?>
+<p>Enter your email address below to unsubscribe:</p>
+<form action="" method="GET">
+    <input type="email" name="a" placeholder="xxxxx@example.com" required />
+    <br />
+    <input type="submit" value="Unsubscribe" />
+</form>
+    <?php
 }
